# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "0.1.9"
    },
    "paths": {
...
        "api/discounted": {
            "post": {
                "tags": [
                    "discounted service"
                ],
                "summary": "create discounted service",
                "description": "Get all request that have been send to a discounted service",
                "parameters": [
                    {
                        "name": "started_date",
                        "in": "query",
                        "description": "quater ",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "end date",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "reduction",
                        "in": "query",
                        "description": "id of user",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "decimal"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "id of city",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "BigInteger"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created"
                    }
                }
            }
        },
        "api/discounted/{discounted}": {
            "patch": {
                "tags": [
                    "discounted service"
                ],
                "summary": "update discounted ",
                "description": "Get all request that have been send to a discounted service",
                "parameters": [
                    {
                        "name": "started_date",
                        "in": "query",
                        "description": "quater ",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "end date",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "reduction",
                        "in": "query",
                        "description": "id of user",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "deciamal"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "id of city",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "BigInteger"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "updated"
                    }
                }
            }
        },
...
        "/api/provider": {
            "get": {
                "tags": [
                    "provider"
                ],
                "summary": "return a list of provider",
                "description": "list of provider",
                "responses": {
                    "200": {
                        "description": "a json array of provider"
                    }
                }
            },
            "post": {
                "tags": [
                    "provider"
                ],
                "summary": "create provider",
                "description": "Get all request that have been send to a provider",
                "parameters": [
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "end date",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "id of provider",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created"
                    }
                }
            }
        },
        "/api/provider{provider}": {
            "delete": {
                "tags": [
                    "provider"
                ],
                "summary": "delete provider by id",
                "description": "delete provider",
                "responses": {
                    "200": {
                        "description": "deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "provider"
                ],
                "summary": "update provider ",
                "description": "Get all request that have been send to a provider",
                "parameters": [
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "end date",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "id of provider",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "updated"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}